requirements.txt
asgiref==3.7.2
astroid==3.0.1
attrs==23.1.0
autopep8==2.0.4
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==5.0
django-allauth==0.58.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
filelock==3.13.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
isort==5.13.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pep8==1.7.1
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.1.0
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
sqlparse==0.4.4
tomlkit==0.12.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
